WIP..adding observe listener
