+++ mobx/lib/mobx.dart
@@ -40,4 +40,5 @@ export 'package:mobx/src/core.dart'
 export 'package:mobx/src/api/action.dart';
 export 'package:mobx/src/api/context.dart';
 export 'package:mobx/src/api/observable.dart';
+export 'package:mobx/src/api/observable_list.dart';
 export 'package:mobx/src/api/reaction.dart';
